% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_aggregate_data.R
\name{aggregate_com_sec}
\alias{aggregate_com_sec}
\title{Aggregate commodities and and sectors in new divisions to reduce the database post-simulation}
\usage{
aggregate_com_sec(
  data = data_full,
  scenarios = c("baseline", unname(scenario)),
  agg_s_table = "aggregation_rules",
  agg_c_table = "aggregation_rules",
  by_com = TRUE,
  by_sec = TRUE,
  bridge_com = bridge_commodities,
  bridge_sec = bridge_sectors,
  exception_s_c = c("CONS", "CONT"),
  detailed.warnings = TRUE
)
}
\arguments{
\item{data}{ThreeME result database}

\item{scenarios}{scenarios existing}

\item{agg_s_table}{which aggregation rules to use for sectors. Default is "aggregation_rules" to use the ones from aggregation_rules.xlsx}

\item{agg_c_table}{which aggregation rules to use for commodities. Default is "aggregation_rules" to use the ones from aggregation_rules.xlsx}

\item{by_com}{Boolean. Whether to aggregate by commodities. Default is TRUE}

\item{by_sec}{Boolean. Whether to aggregate by commodities. Default is TRUE}

\item{bridge_com}{Commodities bridge file  indicating how to aggregate by commodities}

\item{bridge_sec}{Sectors bridge file  indicating how to aggregate by sectors}

\item{exception_s_c}{Character strings of length 4 that start with C or S that do not refer to a sector or commodity}

\item{detailed.warnings}{}
}
\value{
A list of data_full with each version of reaggregated data
}
\description{
Aggregate commodities and and sectors in new divisions to reduce the database post-simulation
}
\keyword{internal}
