% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_catmull.R
\name{catmullrom_splines}
\alias{catmullrom_splines}
\title{Catmull-Rom Cubic splines for interpolation}
\usage{
catmullrom_splines(x_vector, y_vector, steps = 1)
}
\arguments{
\item{x_vector}{Numerical vector of length 4 of the x-axis values (ie dates) to interpolate}

\item{y_vector}{Numerical vector of length 4 of the y-axis values to interpolate}

\item{steps}{numeric (1). Interval between each x values for which to compute the interpolated values. Default: 1.}
}
\value{
named vector of values going from B to C with increments steps. x values are the names of y values of the vector.
}
\description{
Given 4 points A B C D, computes the interpolated values for B-C using Catmull-Rom splines
}
\examples{
catmullrom_splines(sort(sample(1:20,4,FALSE)) , sample(1:30,4,TRUE), steps = 1 )
\dontrun{
 ggplot() +
 geom_point(aes(x= c(7,10,17,20),
              y= c(91,19,70,60)),
          color = "red", size = 3 )+
 geom_point(aes(x = catmullrom_splines(c(7,10,17,20),c(91,19,70,60),steps = 0.1)$x,
                y = catmullrom_splines(c(7,10,17,20),c(91,19,70,60),steps = 0.1)$y  ),
            color = "blue" , size= 1  ) +
 xlab("X")+
 ylab("Y")
}
}
