% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_contrib.plot.R
\name{contrib.plot}
\alias{contrib.plot}
\title{contrib.plot}
\usage{
contrib.plot(
  data,
  series = NULL,
  label_series = NULL,
  line_tot = FALSE,
  startyear = NULL,
  endyear = NULL,
  unit = "percent",
  decimal = 0.1,
  titleplot = "",
  template = template_default
)
}
\arguments{
\item{data}{double(1) a dataframe created with the function loadResults()}

\item{series}{character string, vecteur avec le noms des variables qui seront plottées}

\item{label_series}{character string, vecteur avec le label des variables qui seront plottées}

\item{line_tot}{boolean. Default FALSE . TRUE will display the line for the main variable.}

\item{startyear}{numeric(1), date de début de la période plottée}

\item{endyear}{numeric(1), date de fin  de la période plottée}

\item{unit}{character(1) string , choix de legende pour l'axe des x (en pourcentage ou en niveau)}

\item{decimal}{numeric(1) , choix de la décimale retenue après la virgule}

\item{titleplot}{character(1) string , titre du graphique}

\item{template}{character(1) string, nom du thème ggplot retenu pour le plot}
}
\value{
ggplot
}
\description{
Créer un graphe de type geom_bar à partir d'un vecteur de variables issues d'une base de données
D'autres arguments de la fonction permette de choisir le type d'indicateur et de compléter les labels
}
