% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_contrib.sub.plot.R
\name{contrib.sub.plot}
\alias{contrib.sub.plot}
\title{Contrib.sub.plot}
\usage{
contrib.sub.plot(
  data,
  label_series = NULL,
  startyear = NULL,
  endyear = NULL,
  line_tot = FALSE,
  unit = "percent",
  decimal = 0.1,
  titleplot = "",
  template = template_default
)
}
\arguments{
\item{data}{double(1) a dataframe created with the function loadResults()}

\item{label_series}{character(1) string, vecteur avec le label des variables qui seront tracées}

\item{startyear}{numeric(1), date de début de la période tracée}

\item{endyear}{numeric(1), date de fin de la période tracée}

\item{line_tot}{TRUE pour tracer la courbe correspondant à la variable principale. FALSE par défaut}

\item{unit}{character(1) string , choix de legende pour l'axe des x (en pourcentage ou en niveau)}

\item{decimal}{numeric(1) , choix de la décimale retenue après la virgule}

\item{titleplot}{character(1) string , titre du graphique}

\item{template}{character(1) string, nom du thème ggplot retenu pour le plot}
}
\value{
un ggplot
}
\description{
Créer un graphe de type geom_line à partir d'un vecteur de variables issues d'une base de données
D'autres arguments de la fonction permette de choisir le type d'indicateur et de compléter les labels
}
