% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_contrib.sub.R
\name{contrib.sub}
\alias{contrib.sub}
\title{Contrib.sub computes the contributions of sectors or commodities for a given variable type}
\usage{
contrib.sub(
  data,
  var1,
  group_type = "sector",
  scenar = scenario_to_analyse,
  check_digit = 3
)
}
\arguments{
\item{data}{double(1) a dataframe created with the function loadResults()}

\item{var1}{character(1) string, variable qui seront décomposées en sous-composantes}

\item{group_type}{character(1), type du sous-groupe c("commodity", "sector")}

\item{scenar}{character(1), nom du ou des scénarios}

\item{check_digit}{(1), precision de la vérification}
}
\value{
un dataframe
}
\description{
Contrib.sub computes the contributions of sectors or commodities for a given variable type
}
