% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sector_plots.R
\name{curve_sc_plot}
\alias{curve_sc_plot}
\title{Draw curve plots for a variable decomposed into multiple sectors or commodities}
\usage{
curve_sc_plot(
  data,
  variable,
  group_type = "sector",
  scenario = scenario_name,
  diff = FALSE,
  title = "",
  scenario.names = NULL,
  startyear = NULL,
  endyear = NULL,
  template = template_default,
  scenario.diff.ref = "baseline",
  growth.rate = FALSE,
  abs.diff = FALSE,
  custom_x_breaks = NULL
)
}
\arguments{
\item{data}{a ThreeMe data.frame in long form}

\item{variable}{character string (1) with the root variable to plot. It must be present in the data in both aggregated form and in multiple sector/commodity form. i.e. for output, specify "Y"}

\item{group_type}{length 1 "sector" or "commodity" , "S" or "C" also works}

\item{scenario}{the scenario to plot, must be present in the data}

\item{diff}{If TRUE, the difference from baseline scenario will be plotted. Default is FALSE}

\item{title}{Title of plot to display}

\item{scenario.names}{Name of the scenario to display. If NULL the scenario variable will be used}

\item{startyear}{First year to plot. If NULL (default) reverts to earliest year available.}

\item{endyear}{Last year to plot. If NULL (default) reverts to lastest year available.}

\item{template}{Theme to use (OFCE or nothing)}

\item{scenario.diff.ref}{Name of the reference scenario from which to compute the difference. Default is "baseline"}

\item{growth.rate}{Whether to plot growth rate or not. Default is FALSE (ie will plot levels)}

\item{abs.diff}{If plot is in difference TRUE will plot absolute difference, FALSE (default) will plot relative difference}

\item{custom_x_breaks}{integer(1) permet de choisir manuellement les ticks des années à afficher. Si NULL (défaut) alors utilise l'algorithme de la fonction. Si "R" : algorithme par défaut de R sera utilisé}
}
\value{
a ggplot
}
\description{
Draw curve plots for a variable decomposed into multiple sectors or commodities
}
