% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_custom_palette.R
\name{custom.palette}
\alias{custom.palette}
\title{custom_palette}
\usage{
custom.palette(
  n = NULL,
  bridge = NULL,
  sub.bridge = FALSE,
  palette_base = NULL
)
}
\arguments{
\item{n}{numeric(1) A number of colors to pick}

\item{bridge}{If not null, ThreeME bridge to be loaded takes precedent over n}

\item{sub.bridge}{If TRUE, the complete list of sectors is chosen when defined by aggregated sectors, by default FALSE}

\item{palette_base}{If not null, a vector of colors to based the palette}
}
\value{
a vector of colors (hex code)
}
\description{
custom_palette
}
\examples{
custom.palette(n = 6)
}
