% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_to_qmd.R
\name{decompose_tex}
\alias{decompose_tex}
\title{Decompose equation tex file}
\usage{
decompose_tex(
  preface_file = "03.1-eq_preface.tex",
  maintex_file = "03.1-eq.tex"
)
}
\arguments{
\item{preface_file}{path to preface file}

\item{maintex_file}{path to main tex file}
}
\value{
data.frame list of data frames with lines treated from the texfiles
}
\description{
Decompose equation tex file
}
