% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviewsmodelsolver.R
\name{eviews_model_solver}
\alias{eviews_model_solver}
\title{Call eviews solver}
\usage{
eviews_model_solver(
  config_file = configuration,
  before_solving_data = NULL,
  overwrite_eviews = NULL,
  limit_size_calib_csv = 12
)
}
\arguments{
\item{config_file}{Ermeeth configuration list}

\item{before_solving_data}{database before data is solved}

\item{overwrite_eviews}{path to eviews.exe}

\item{limit_size_calib_csv}{import stringr}
}
\description{
Call eviews solver
}
