% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_theme.R
\name{flextable_theme}
\alias{flextable_theme}
\title{flextable_theme used in standard shocs}
\usage{
flextable_theme(
  x,
  colnumber = if (exists("col_numb")) {
     col_numb
 } else {
     6
 },
  vline = c(2),
  palette_confetti = NULL,
  donotcolor = 1,
  scales = TRUE,
  minimum = -3,
  maximum = 3,
  fontsize = 12
)
}
\arguments{
\item{x}{A flextable table}

\item{colnumber}{number of columns}

\item{vline}{vector of columns where to draw thicker lines}

\item{palette_confetti}{Palette to use, made by the confetti function}

\item{donotcolor}{vector of columns where not to colour}

\item{scales}{Boolean, whether or not to use the red-green colours scale. Default: TRUE}

\item{minimum}{minimum of colour scale}

\item{maximum}{maximum of colour scale}

\item{fontsize}{size of font in the table. Default: 12}
}
\value{
a flextable with the format
}
\description{
flextable_theme used in standard shocs
}
