% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_shared_3me.R
\name{get_file_structure}
\alias{get_file_structure}
\title{File structure of the shared threeme data}
\usage{
get_file_structure(
  url = "https://raw.githubusercontent.com/ThreeME-org/shared_3me_data/main/tree.txt"
)
}
\arguments{
\item{url}{location of the tree.txt file}
}
\value{
dataframe with folder contents
}
\description{
File structure of the shared threeme data
}
