% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_shared_3me.R
\name{get_qmd_template}
\alias{get_qmd_template}
\title{Pre programmed get_remote_file function for qmd templates}
\usage{
get_qmd_template(...)
}
\arguments{
\item{...}{arguments to be passed on to get_remote_file function}
}
\value{
downloads file
}
\description{
Pre programmed get_remote_file function for qmd templates
}
