% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_regex}
\alias{get_regex}
\title{get all elements from a vector with multiple regexes}
\usage{
get_regex(
  regex_list = c("*"),
  vector = c(""),
  negate = FALSE,
  case_sensitive = FALSE,
  results_as_list = FALSE
)
}
\arguments{
\item{regex_list}{character vector of regexes to look for in a vector}

\item{vector}{character vector where to look for}

\item{negate}{boolean, set to TRUE to extract elements that are DO NOT verify the regexes. Default : FALSE}

\item{case_sensitive}{boolean, set to FALSE to ignore the case from your search. Default : FALSE}

\item{results_as_list}{boolean, set to TRUE if looking for results in list form, where every element is a character vector of elements from the search vector verifying (or not if negate = FALSE) the regexes.}
}
\value{
character vector compiled
}
\description{
get all elements from a vector with multiple regexes
}
\examples{
get_regex(regex_list = c("y$"), vector = months)

}
