% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sec_com.R
\name{get_sec_com}
\alias{get_sec_com}
\title{Get the sector and commodities codes from an ermeeth}
\usage{
get_sec_com(
  list.mdl = file.path("src", "model", "threeme", paste0("R_lists_", config$input$iso3,
    "_", config$input$classification, ".mdl")),
  sectors = TRUE,
  commodities = TRUE,
  energy_commodities = TRUE
)
}
\arguments{
\item{list.mdl}{file path to an mdl list file usually found in the src/model folder}

\item{sectors}{boolean, to return the sector list, will return a NULL element otherwise. Default is TRUE.}

\item{commodities}{boolean, to return the commodity list, will return a NULL element otherwise Default is TRUE.}

\item{energy_commodities}{boolean, to return the energy_commodity list, will return a NULL element otherwise Default is TRUE.}
}
\value{
a list of vectors for sectors, commodities and energy commodities codes
}
\description{
Get the sector and commodities codes from an ermeeth
}
\examples{
\dontrun{
get_sec_com(list.mdl = file.path("src","model","threeme","R_lists_FRA_c29_s33.mdl"))
}

}
