% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4a_dynamo_install.R
\name{install_dynamo}
\alias{install_dynamo}
\title{Install Dynamo}
\usage{
install_dynamo(dynamo_path = getwd(), overwrite = TRUE)
}
\arguments{
\item{dynamo_path}{directory path where to install Dynamo}

\item{overwrite}{TRUE to overwrite existing Dynamo install files}
}
\value{
Dynamo installation in the specified location
}
\description{
This function installs the Dynamo executable in a specified location
}
\examples{
\dontrun{
 install_dynamo("../src/compiler")
}
}
