% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_interpolation.R
\name{interpolation_series}
\alias{interpolation_series}
\title{interpolation_series}
\usage{
interpolation_series(
  date_vector = NULL,
  value_vector = NULL,
  method = "linear",
  seq.step = 1,
  first.date = first_date,
  last.date = last_date,
  showplot = FALSE
)
}
\arguments{
\item{date_vector}{numeric or Date vector to match with value vector}

\item{value_vector}{numeric vector of anchor values from which to compute interpolation}

\item{method}{"linear" only for now}

\item{seq.step}{regular interval expected between each date}

\item{first.date}{date from which to start the final sequence.}

\item{last.date}{date where the final sequence stops}

\item{showplot}{show the plot of the computations}
}
\value{
a numeric vector
}
\description{
Given a date vector and corresponding value vector, computes the values in between each date using some form of interpolation .
}
