% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Title}
\usage{
label(variable_code, data = ermeeth::label_database, lang = language)
}
\arguments{
\item{variable_code}{a character or a character vector}

\item{data}{a R data frame}

\item{lang}{label destination language}
}
\value{
a character or a character vector
}
\description{
Title
}
\examples{
\dontrun{label("GDP")}
}
