% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b_long_data.R
\name{long_data}
\alias{long_data}
\title{Convert back a wide ThreeMe data.frame into a long ThreeMe data.frame}
\usage{
long_data(
  data,
  sector_names_table = names_sectors,
  commodity_names_table = names_commodities
)
}
\arguments{
\item{data}{Wide ThreeMe data.frame or list of wide ThreeMe data.frame as created by the wide_data function}

\item{sector_names_table}{names table for sectors as found in codenames_xxx_xxx.R files. if NULL, sector names will rename coded}

\item{commodity_names_table}{names table for commodity as found in codenames_xxx_xxx.R files. if NULL, commodity names will rename coded}
}
\value{
a long form ThreeMe dataframe
}
\description{
Convert back a wide ThreeMe data.frame into a long ThreeMe data.frame
}
