% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_to_qmd.R
\name{make_eq_qmd}
\alias{make_eq_qmd}
\title{Transform the tex equation files into 3 qmds}
\usage{
make_eq_qmd(
  preface = "03.1-eq_preface.tex",
  maintex = "03.1-eq.tex",
  path = file.path("results", "quarto_templates", "results_side_files"),
  out.dir = file.path("results", "quarto_templates", "results_side_files")
)
}
\arguments{
\item{preface}{path to preface tex file created by teXdoc function}

\item{maintex}{path to maintex tex file created by teXdoc function}

\item{path}{files path}

\item{out.dir}{output directory}
}
\value{
3 or less qmd files
}
\description{
Transform the tex equation files into 3 qmds
}
