% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_3me}
\alias{message_3me}
\title{Basic formattinf function for display messages}
\usage{
message_3me(
  text_message,
  colour_bg = "seagreen2",
  text_black = TRUE,
  symbol_u = "✅"
)
}
\arguments{
\item{text_message}{character string for the text to display}

\item{colour_bg}{character string colour of the background}

\item{text_black}{boolean, TRUE to keep the text in black, FALSE to make it white}

\item{symbol_u}{unicode symbol to display before the message}
}
\value{
cat message with formatting
}
\description{
Basic formattinf function for display messages
}
