% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_any}
\alias{message_any}
\title{Preprogrammed ThreeME formatting for normal messages, with possibility to customize the symbol}
\usage{
message_any(txt_message = "TEST", custom_symbol = "")
}
\arguments{
\item{txt_message}{text to display}

\item{custom_symbol}{custom unicode symbol to display, Default displays none}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for normal messages, with possibility to customize the symbol
}
