% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_delete}
\alias{message_delete}
\title{Preprogrammed ThreeME formatting for Deleting messages}
\usage{
message_delete(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for Deleting messages
}
