% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_main_step}
\alias{message_main_step}
\title{Preprogrammed ThreeME formatting for MAIN STEPS messages}
\usage{
message_main_step(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for MAIN STEPS messages
}
