% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_not_ok}
\alias{message_not_ok}
\title{Preprogrammed ThreeME formatting for NOT Ok messages}
\usage{
message_not_ok(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for NOT Ok messages
}
