% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_save}
\alias{message_save}
\title{Preprogrammed ThreeME formatting for SAVING messages}
\usage{
message_save(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for SAVING messages
}
