% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_stopbomb}
\alias{message_stopbomb}
\title{Preprogrammed ThreeME formatting for STOP messages}
\usage{
message_stopbomb(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for STOP messages
}
