% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_display_messages.R
\name{message_warning}
\alias{message_warning}
\title{Preprogrammed ThreeME formatting for warning messages}
\usage{
message_warning(txt_message = "TEST")
}
\arguments{
\item{txt_message}{text to display}
}
\value{
cat messages
}
\description{
Preprogrammed ThreeME formatting for warning messages
}
