% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_export.R
\docType{data}
\name{oilprice_fra_agg}
\alias{oilprice_fra_agg}
\title{oilprice_fra_agg}
\format{
data.frame
}
\source{
ThreeMe simulations
}
\usage{
data(oilprice_fra_agg)

##Dataexport
}
\description{
full ThreeMe database with sector and commodity aggregation
}
\author{
ThreeMe team
}
\keyword{data}
