% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_small_functions.R
\name{pythagore}
\alias{pythagore}
\title{Theoreme de pythagore}
\usage{
pythagore(a = 1, b = 5, fr_format = FALSE)
}
\arguments{
\item{a}{numeric (length 1), the first side of the triangle. Default 1}

\item{b}{numeric (length 1), the second side of the triangle. Default 1}

\item{fr_format}{afficher format francais}
}
\value{
character length 1 , hypothenuse
}
\description{
Cette fonction calcul l'hypothénuse d'un triangle rectangle
}
\examples{
pythagore(a = 4, b = 3)
}
