% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_quadratic_splines.R
\name{quadratric_splines}
\alias{quadratric_splines}
\title{Quadratic splines for interpolation}
\usage{
quadratric_splines(x_vector, y_vector, side = "left", steps = 1)
}
\arguments{
\item{x_vector}{Numerical vector of length 3 of the x-axis values (ie dates) to interpolate}

\item{y_vector}{Numerical vector of length 3 of the y-axis values to interpolate}

\item{side}{Either "left" to compute A-B interpolation or "right" to compute B-C interpolation. Default: "left"}

\item{steps}{numeric (1). Interval between each x values for which to compute the interpolated values. Default: 1.}
}
\value{
named vector of values going from B to C with increments steps. x values are the names of y values of the vector.
}
\description{
Given 3 points A, B and C computes the interpolated values for A-B or B-C using quadratic splines
}
\examples{
quadratric_splines(c(9,10,14),c(91,19,70),steps = 0.5, side= "left")
quadratric_splines(c(9,10,14),c(91,19,70),steps = 0.5, side= "right")
\dontrun{
ggplot(mapping = aes(x= c(9,10,14), y= c(91,19,70))) +
geom_point(color = "red", size = 3)+
geom_point(aes(x = quadratric_splines(c(9,10,14),c(91,19,70),steps = 0.1, side= "left")$x,
                y = quadratric_splines(c(9,10,14),c(91,19,70),steps = 0.1, side= "left")$y),
                 color = "blue" , size= 1  )+

 geom_point(aes(x = quadratric_splines(c(9,10,14),c(91,19,70),steps = 0.1, side= "right")$x,
                y = quadratric_splines(c(9,10,14),c(91,19,70),steps = 0.1, side= "right")$y),
                color = "purple" , size= 1  ) +
 xlab("X")+
 ylab("Y")
 }
}
