% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data_load.R
\name{read_3me_eviews_csv}
\alias{read_3me_eviews_csv}
\title{Read the eviews csv output from simulations}
\usage{
read_3me_eviews_csv(
  csv.file.path = file.path("csv", "ct1.csv"),
  variables_selection = NULL,
  scenario_label = NULL
)
}
\arguments{
\item{csv.file.path}{file path to the EViews simulation csv output}

\item{variables_selection}{vector of variables}

\item{scenario_label}{}
}
\value{
a data.frame containing the simulation output
}
\description{
Read the eviews csv output from simulations
}
