% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_small_functions.R
\name{redirect}
\alias{redirect}
\title{redirect
when you need to change the directed path because of markdowns}
\usage{
redirect(normal_path, redirect = path_main)
}
\arguments{
\item{normal_path}{normal path of the file}

\item{redirect}{new path of the file}
}
\value{
character string with the path changed
}
\description{
redirect
when you need to change the directed path because of markdowns
}
