% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b_runDynamo.R
\name{runDynaMo}
\alias{runDynaMo}
\title{Run Dynamo Compiler for models}
\usage{
runDynaMo(
  iso3,
  baseyear,
  lastyear,
  calib,
  model,
  max_lags = 3,
  dynamo_path = file.path("src")
)
}
\arguments{
\item{iso3}{country code extension}

\item{baseyear}{baseyear of simulations}

\item{lastyear}{last year of simulations}

\item{calib}{vector list of calibration files}

\item{model}{vector list of model files}

\item{max_lags}{number of years before baseyear to generate based on lags used the equations of the model}

\item{dynamo_path}{path where dynamo is with the calib files and model files}
}
\value{
created the model.prg and calib.csv files
}
\description{
Run Dynamo Compiler for models
}
