% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_selectseries.R
\name{selectseries}
\alias{selectseries}
\title{selectseries extract series from a long form data.frame and apply a transformation}
\usage{
selectseries(
  data,
  scenarios,
  series,
  startyear = NULL,
  endyear = NULL,
  transformation = "reldiff",
  baseline_name = "baseline"
)
}
\arguments{
\item{data}{a long form ThreeMe data.frame generated by loadResults.}

\item{scenarios}{the name of the scenario (must match the name of scenario column in the data.frame). Length : 1}

\item{series}{the series to select and transform}

\item{startyear}{First year of the sample to extract. If NULL (default) the earliest year available will be used.}

\item{endyear}{Last year of the sample to extract. If NULL (default) the lastest year available will be used.}

\item{transformation}{The transformation to apply. The options are : "reldiff" (default) for relative difference with the baseline, "diff" for the absolute difference with the baseline. Anything else will plot levels}

\item{baseline_name}{Name of the baseline scenario, Default: "baseline"}
}
\value{
a long form data.frame with only the selected series and years and new columns with some calculations over the scenarios.
}
\description{
selectseries extract series from a long form data.frame and apply a transformation
}
