% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_simpleplot.R
\name{simpleplot}
\alias{simpleplot}
\title{Simple plot}
\usage{
simpleplot(
  data,
  series,
  label_series = NULL,
  startyear = NULL,
  endyear = NULL,
  transformation = "reldiff",
  unit = "percent",
  decimal = 0.1,
  titleplot = "",
  scenario = scenario_name,
  template = template_default
)
}
\arguments{
\item{data}{double(1) a dataframe created with the function loadResults()}

\item{series}{character(1) string, vecteur avec le noms des variables qui seront plottées}

\item{label_series}{character(1) string, vecteur avec le label des variables qui seront plottées}

\item{startyear}{numeric(1), date de début de la période plottée}

\item{endyear}{numeric(1), date de fin  de la période plottée}

\item{transformation}{character(1) string, choix de l'option de l'indicateur (en niveau, en variation, différence absolue, relative )}

\item{unit}{character(1) string , choix de legende pour l'axe des x (en pourcentage ou en niveau)}

\item{decimal}{numeric(1) , choix de la décimale retenue après la virgule}

\item{titleplot}{character(1) string , titre du graphique}

\item{scenario}{character(1) string, nom du scénario à plotter, doit être dans la base de données.}

\item{template}{character(1) string, nom du thème ggplot retenu pour le plot}
}
\value{
un ggplot
}
\description{
Créer un graphe de type geom_line à partir d'un vecteur de variables issues d'une base de données
D'autres arguments de la fonction permette de choisir le type d'indicateur et de compléter les labels
}
