% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_sector_plots.R
\name{stacked_sc_plot}
\alias{stacked_sc_plot}
\title{Draw stacked bar plots  for a variable decomposed into multiple sectors or commodities}
\usage{
stacked_sc_plot(
  data,
  variable,
  group_type = "sector",
  scenario = scenario_name,
  diff = FALSE,
  title = "",
  scenario.names = NULL,
  interval = 10,
  start_year = NULL,
  end_year = NULL,
  scenario.diff.ref = "baseline",
  template = template_default,
  corner_text = "in Millions"
)
}
\arguments{
\item{data}{a ThreeMe data.frame in long form}

\item{variable}{character string (1) with the root variable to plot. It must be present in the data in both aggregated form and in multiple sector/commodity form. i.e. for output, specify "Y"}

\item{group_type}{length 1 "sector" or "commodity" , "S" or "C" also works}

\item{scenario}{the scenario to plot, must be present in the data}

\item{diff}{If TRUE, the difference from baseline scenario will be plotted. Default is FALSE}

\item{title}{Title of plot to display}

\item{scenario.names}{Name of the scenario to display. If NULL the scenario variable will be used}

\item{interval}{interval of years to plot}

\item{start_year}{First year to plot. If NULL (default) reverts to earliest year available.}

\item{end_year}{Last year to plot. If NULL (default) reverts to lastest year available.}

\item{scenario.diff.ref}{Name of the reference scenario from which to compute the difference. Default is "baseline"}

\item{template}{Theme to use (OFCE or nothing)}

\item{corner_text}{y-axis legend to appear in top left corner}
}
\value{
a ggplot
}
\description{
Draw stacked bar plots  for a variable decomposed into multiple sectors or commodities
}
