% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_table.output.R
\name{table.output}
\alias{table.output}
\title{Table.output}
\usage{
table.output(
  data = data,
  scenario = scenario_name,
  startyear = NULL,
  endyear = NULL,
  export.doc = TRUE,
  langue = "en",
  full.table = TRUE,
  title = NULL,
  decimal = 2,
  year.index = NULL,
  results.folder = getwd()
)
}
\arguments{
\item{data}{double(1) a dataframe created with the function loadResults()}

\item{scenario}{character(1) string, vecteur avec le label des variables qui seront tracées}

\item{startyear}{startyear for the table, by default, startyear of the dataframe}

\item{endyear}{endyear for the table, by default, endyear of the dataframe}

\item{export.doc}{boolean if TRUE, export a docx into the folder results}

\item{langue}{character(1) string to specify the language of comments c('en', 'fr)}

\item{full.table}{boolean if TRUE, include other macroeconomic indicators ,TRUE by default}

\item{title}{character(1) string of character for the title of the table}

\item{decimal}{numeric : number of decimal in the datable}

\item{year.index}{boolean if TRUE, use years for columns, if FALSE number of periods with respect to first date}

\item{results.folder}{character(1): pathfile to export the results. By defaut the program folder}
}
\value{
a flextable
}
\description{
Table.output
}
\examples{
\dontrun{
table.output(data = data, scenario = "oilprice_fra",
full.table = TRUE, export.doc = FALSE, title = 'Gros test')
}
}
