% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_texdoc.R
\name{teXdoc}
\alias{teXdoc}
\title{teXdoc takes a model file}
\usage{
teXdoc(
  sources,
  exo = c(),
  base.path = "src/model",
  out = "doc",
  out.path = getwd(),
  compile_pdf = TRUE
)
}
\arguments{
\item{sources}{mdl source files containing the equations}

\item{exo}{mdl source files containing the exogenous variables}

\item{base.path}{source files directory}

\item{out}{output file names (no directory needed, the output directory is to be specified in out.path)}

\item{out.path}{output files directory (default is working directory)}

\item{compile_pdf}{TRUE (default) will compile the tex files into a pdf}
}
\value{
creates the necessary tex files to compile the model's documentation and compiles a pdf version
}
\description{
teXdoc takes a model file
}
\examples{
\dontrun{
 teXdoc(sources   = c("model_test.mdl"),
        exo       = c("model_test-exovar.mdl"),
        base.path = system.file(package = "ermeeth"),
        out       = "documentation-eq",
        out.path  = "documentation_test")
}
}
