% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_shared_3me.R
\name{threeme_scenario_labels}
\alias{threeme_scenario_labels}
\title{Standard ThreeME scenario labels}
\usage{
threeme_scenario_labels()
}
\value{
named vector containing labels for each scenario
}
\description{
Standard ThreeME scenario labels
}
