% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad.R
\name{trad}
\alias{trad}
\title{Title}
\usage{
trad(
  x,
  trad_data_base = trad_language_base,
  lang = language,
  lang_source = "en"
)
}
\arguments{
\item{x}{a character}

\item{trad_data_base}{a R data frame}

\item{lang}{translation destination language}

\item{lang_source}{translation source language}
}
\value{
a character vector
}
\description{
Title
}
\examples{
\dontrun{trad("Alternatives", lang = "fr")}
}
