% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_small_functions.R
\name{variables_like}
\alias{variables_like}
\title{Title}
\usage{
variables_like(data, test, view = TRUE)
}
\arguments{
\item{data}{A ThreeMe data.frame in long format.}

\item{test}{A regular expression to test for matching variables.}

\item{view}{If TRUE (default) results will be displayed in a separate window.}
}
\value{
A data.frame containing the variables matching the test expression, as well as sectors and commodities corresponding to these variables where applicable.
}
\description{
Title
}
