% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_wide_data.R
\name{wide_data}
\alias{wide_data}
\title{Transform a long ThreeMe database to wide format}
\usage{
wide_data(
  data,
  scenarios = c("baseline", scenario_name),
  variables = NULL,
  out_format = "dataframe"
)
}
\arguments{
\item{data}{data.frame: A ThreeMe dataframe (long format)}

\item{scenarios}{character vector: the scenarios that should be selected, by default uses baseline and scenario_name}

\item{variables}{character vector: the variables to be selected. NULL by default, takes all the variables in the database}

\item{out_format}{output format, either 'dataframe' or 'list'. By default 'dataframe' returns a dataframe with the name of scenario preceding variable name}
}
\value{
a data.frame or a list
}
\description{
Transform a long ThreeMe database to wide format
}
