% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_interpolation.R
\name{interpolation_series}
\alias{interpolation_series}
\title{interpolation_series}
\usage{
interpolation_series(
  date_vector = NULL,
  value_vector = NULL,
  smoothed = TRUE,
  seq.step = 1,
  first.date = first_date,
  last.date = last_date,
  full_results = FALSE,
  debug = FALSE
)
}
\arguments{
\item{date_vector}{numeric or Date vector to match with value vector}

\item{value_vector}{numeric vector of anchor values from which to compute interpolation}

\item{smoothed}{If FALSE, interpolation will be on done only linearly. If TRUE : If 3 observations : quadratic splines will be used. If 4 or more observations : cubic splines via Catmull-Rom method will be be used. If 2 observations only : reverts to linear interpolation. Number of observations include additional observations added to reach first.date and last.date. Default: TRUE}

\item{seq.step}{numeric (1). Interval between each x values for which to compute the interpolated values. Default: 1.}

\item{first.date}{Date from which to start the final sequence. Default: ThreeME "first_date" argument.}

\item{last.date}{Date where the final sequence stops. Default: ThreeME "last_date" argument.}

\item{full_results}{If TRUE a list containing the input, the resulting interpolation by two methods, the results as a named vector and the plot will be returned. Otherwise only  the named vector presenting the resulting values will be returned. Default: FALSE}

\item{debug}{If TRUE, will enter debug mode. Default: FALSE.}
}
\value{
Either a named vector of values going from first.date to last.date with the chosen interpolation method (smoothed or not), or a list with multiple objects including the plot and a data.frame with linear and smoothed interpolation
}
\description{
Given a date vector and corresponding value vector, computes the values in between each date using some form of interpolation.
}
\examples{
interpolation_series(date_vector = c(2001, 2005, 2010, 2015, 2020),
value_vector = c(20,30,10,25,45), first.date = 2000, last.date = 2020, full_results = TRUE)

}
