% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_interpolation.R
\name{linear_fill}
\alias{linear_fill}
\title{linear_fill}
\usage{
linear_fill(vector)
}
\arguments{
\item{vector}{a numeric vector}
}
\value{
a numeric vector of the same length as vector
}
\description{
Takes a vector and computes the 2nd to the penultimate values using the first and last value of the input vector using linear interpolation, assuming each element of the vector are regularly spaced out.
}
\examples{
linear_fill(c(1,100,2))
linear_fill(c(42,NA,NA,NA,32))
}
