% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_load_excel_calib.R
\name{load_excel_calibration}
\alias{load_excel_calibration}
\title{Load excel calibration for ThreeME and fill in the missing values through spline interpolation}
\usage{
load_excel_calibration(
  excel_sheet = "configuration/scenarii_calib/scenarii_inputs.xlsx",
  sheet_to_load = "baseline",
  calib_base = "src/compiler/calib.csv",
  base_year = baseyear,
  check_tol = 1e-07,
  stop_if_calib_fail = TRUE,
  keep_baseyear_calib_data = TRUE
)
}
\arguments{
\item{excel_sheet}{character(1) path to the excel sheet with the calib data}

\item{sheet_to_load}{sheet to load for the calibration}

\item{calib_base}{calib.csv in threeME R location (default is "src/compiler/calib.csv", unlikely to be changed)}

\item{base_year}{base year of the model}

\item{check_tol}{tolerance check for interpolation verification}

\item{stop_if_calib_fail}{abort if interpolation results do not match calibration}

\item{keep_baseyear_calib_data}{if missmatch keep the calibration year data}
}
\value{
a data frame with the series from excel with interpolation if needed
}
\description{
Load excel calibration for ThreeME and fill in the missing values through spline interpolation
}
