% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_loadEviewscsv.R
\name{read_3me_eviews_csv}
\alias{read_3me_eviews_csv}
\title{Read ThreeME simulation csv output generated by EViews}
\usage{
read_3me_eviews_csv(
  csv.file.path = file.path("csv", "ct1.csv"),
  variables_selection = variables_to_keep,
  scenario_label = NULL
)
}
\arguments{
\item{csv.file.path}{file.path where to find the csv file}

\item{variables_selection}{optional in order to import only certain variables}

\item{scenario_label}{name of shock scenario to import, if NULL will use the csv filename}
}
\value{
a long format threeme dataframe
}
\description{
Read ThreeME simulation csv output generated by EViews
}
