% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trad.R
\name{trad}
\alias{trad}
\title{Traduction à partir d'une base de données de langue}
\usage{
trad(x, data = trad_database, lang = language, lang_source = "en")
}
\arguments{
\item{x}{a character}

\item{data}{a R data frame}

\item{lang}{translation destination language}

\item{lang_source}{translation source language}
}
\value{
a character vector
}
\description{
Traduction à partir d'une base de données de langue
}
\examples{
\dontrun{trad("Alternatives", lang = "fr")}
}
